+++ internal/mining/util/set.go
