+++ internal/mining/util/timeout.go
